Fixed ProxyWalletInstance Error
Problem was web3.eth.Contract() method was expecting a string
for address. Type cast the object from map to string solved this.

Brownie has an odd error that will think exchangeRateCurrent()
is a re-entrancy if run for the first time. Need to run
exchangeRateCurrent() as a transaction in brownie console first.