Check and Deploy Clone for User
we check to see if the user has deployed a clone with getClon(). If it throws an error,
we ask the user to send a transaction to getOrCreateClone(). We can then get the clone
address from getClone() once the clone is created.

Since we have the cloneAddress, we can create a web3 contract object