Fixed error in helper from merge, added future expiry check
There was an error in helper from the merge. Fixed that.
Added a future Expiry check - if the helper script runs right when the block height
is around a multiple of 4096, then it may create an expiry, but when the webapp looks
for the expiry, it may be trying to find the next 'chunk'. So there is an if statement
to say 'check the prior chunk'.